+++ packages/pointer_interceptor/pointer_interceptor/example/.metadata
@@ -4,7 +4,27 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: e6bd95bc5caa5e34c5b0285a559673984374b7ea
-  channel: master
+  revision: "61a388a44849d19d0b8586e6650ebae901af75d5"
+  channel: "main"
 
 project_type: app
+
+# Tracks metadata for the flutter migrate command
+migration:
+  platforms:
+    - platform: root
+      create_revision: 61a388a44849d19d0b8586e6650ebae901af75d5
+      base_revision: 61a388a44849d19d0b8586e6650ebae901af75d5
+    - platform: ios
+      create_revision: 61a388a44849d19d0b8586e6650ebae901af75d5
+      base_revision: 61a388a44849d19d0b8586e6650ebae901af75d5
+
+  # User provided section
+
+  # List of Local paths (relative to this file) that should be
+  # ignored by the migrate tool.
+  #
+  # Files that are not part of the templates will be ignored by default.
+  unmanaged_files:
+    - 'lib/main.dart'
+    - 'ios/Runner.xcodeproj/project.pbxproj'
