+++ packages/pointer_interceptor/pointer_interceptor/example/ios/Flutter/Debug.xcconfig
