+++ packages/pointer_interceptor/pointer_interceptor/example/ios/Runner/Runner-Bridging-Header.h
@@ -1,0 +1,1 @@
+#import "GeneratedPluginRegistrant.h"
