+++ packages/pointer_interceptor/pointer_interceptor/example/lib/main.dart
@@ -2,50 +2,20 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// ignore: avoid_web_libraries_in_flutter
-import 'dart:html' as html;
-import 'dart:ui_web' as ui_web;
-
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:pointer_interceptor/pointer_interceptor.dart';
+import 'package:pointer_interceptor_example/example_platform_view.dart';
 
-const String _htmlElementViewType = '_htmlElementViewType';
+import 'example_platform_view.dart' if (dart.library.html) 'html_element.dart' as html;
+
 const double _videoWidth = 640;
 const double _videoHeight = 480;
 
-/// The html.Element that will be rendered underneath the flutter UI.
-html.Element htmlElement = html.DivElement()
-  ..style.width = '100%'
-  ..style.height = '100%'
-  ..style.backgroundColor = '#fabada'
-  ..style.cursor = 'auto'
-  ..id = 'background-html-view';
-
-// See other examples commented out below...
-
-// html.Element htmlElement = html.VideoElement()
-//   ..style.width = '100%'
-//   ..style.height = '100%'
-//   ..style.cursor = 'auto'
-//   ..style.backgroundColor = 'black'
-//   ..id = 'background-html-view'
-//   ..src = 'https://archive.org/download/BigBuckBunny_124/Content/big_buck_bunny_720p_surround.mp4'
-//   ..poster = 'https://peach.blender.org/wp-content/uploads/title_anouncement.jpg?x11217'
-//   ..controls = true;
-
-// html.Element htmlElement = html.IFrameElement()
-//       ..width = '100%'
-//       ..height = '100%'
-//       ..id = 'background-html-view'
-//       ..src = 'https://www.youtube.com/embed/IyFZznAk69U'
-//       ..style.border = 'none';
-
 void main() {
-  ui_web.platformViewRegistry.registerViewFactory(
-    _htmlElementViewType,
-    (int viewId) => htmlElement,
-  );
-
+  // if (kIsWeb) {
+  //   registerWebPlatformView();
+  // }
   runApp(const MyApp());
 }
 
@@ -114,12 +84,14 @@ class _MyHomePageState extends State<MyHomePage> {
               child: Stack(
                 alignment: Alignment.center,
                 children: <Widget>[
-                  HtmlElement(
-                    key: const ValueKey<String>('background-widget'),
-                    onClick: () {
-                      _clickedOn('html-element');
-                    },
-                  ),
+                  // kIsWeb?
+                  //   HtmlElement(
+                  //     key: const ValueKey<String>('background-widget'),
+                  //     onClick: () {
+                  //       _clickedOn('html-element');
+                  //     },
+                  //   ) :
+                    ExamplePlatformView(),
                   Row(
                     mainAxisAlignment: MainAxisAlignment.spaceEvenly,
                     children: <Widget>[
@@ -194,26 +166,6 @@ class _MyHomePageState extends State<MyHomePage> {
           ),
         ),
       ),
-    );
-  }
-}
-
-/// Initialize the videoPlayer, then render the corresponding view...
-class HtmlElement extends StatelessWidget {
-  /// Constructor
-  const HtmlElement({super.key, required this.onClick});
-
-  /// A function to run when the element is clicked
-  final VoidCallback onClick;
-
-  @override
-  Widget build(BuildContext context) {
-    htmlElement.onClick.listen((_) {
-      onClick();
-    });
-
-    return const HtmlElementView(
-      viewType: _htmlElementViewType,
     );
   }
 }
