+++ packages/pointer_interceptor/pointer_interceptor/lib/pointer_interceptor.dart
@@ -7,7 +7,6 @@ library pointer_interceptor;
 import 'package:flutter/widgets.dart';
 import 'package:pointer_interceptor_platform_interface/pointer_interceptor_platform_interface.dart';
 
-export 'package:pointer_interceptor_platform_interface/pointer_interceptor_platform_interface.dart';
-export 'src/web.dart';
+// export 'src/web.dart';
 
 part 'src/pointer_interceptor_widget.dart';
