+++ packages/pointer_interceptor/pointer_interceptor_ios/lib/pointer_interceptor_ios.dart
@@ -1,8 +1,23 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
 
-import 'pointer_interceptor_ios_platform_interface.dart';
+import 'package:flutter/widgets.dart';
+import 'package:pointer_interceptor_platform_interface/pointer_interceptor_platform_interface.dart';
 
-class PointerInterceptorIos {
-  Future<String?> getPlatformVersion() {
-    return PointerInterceptorIosPlatform.instance.getPlatformVersion();
+/// The iOS implementation of the [PointerInterceptorPlatform]
+class PointerInterceptorPlugin extends PointerInterceptorPlatform {
+
+  static void registerWith() {
+    PointerInterceptorPlatform.instance = PointerInterceptorPlugin();
+  }
+
+  @override
+  Widget buildWidget(
+      {required Widget child,
+      bool intercepting = true,
+      bool debug = false,
+      Key? key}) {
+    return child;
   }
 }
